internal: refactor array methods into separate private submodule
